correzioni
aggiustamenti su crud e...
boh i template erano tutti corretti anche mi saltava in aria
modificandoli sono riuscito a vederli, credo fosse un errore del motore

+++ pages/admin/customers/delete.php
@@ -1,4 +1,5 @@
 <?php
+
 include '../../../classes/dataBase.php';
 
 $id = $_GET['id'];
@@ -8,27 +9,21 @@ try {
     $db = new dataBase();
     $DBH = $db->connect();
 
-    //cerco categorie associate
-//    $stmt = $DBH->prepare('SELECT * FROM categories WHERE categories_id = :id');
-//    $stmt->execute($data);
-//    $categories = $stmt->fetchAll();
-//
-//    if (empty($categories)) {
-//        //cerco prodotti associati
-//        $STH = $DBH->prepare('SELECT * FROM products WHERE categories_id = :id');
-//        $STH->execute($data);
-//        $products = $STH->fetchAll();
-//        if (empty($products)) {
-            //se non ho vincoli, elimino.
-            $STH = $DBH->prepare('DELETE FROM customers WHERE id = :id');
-            $STH->execute($data);
-            $message = 'Delete successful';
-//        } else {
-//            $message = 'Cannot delete because of depency with associate products';
-//        }
-//    } else {
-//        $message = 'Cannot delete because of depency with another category';
-//    }
+    //cerco ordini associati
+    $stmt = $DBH->prepare('SELECT * FROM orders WHERE customers_id = :id');
+    $stmt->execute($data);
+    $orderss = $stmt->fetchAll();
+
+    if (empty($orders)) {
+        //se non ho vincoli, elimino.
+        $STH = $DBH->prepare('DELETE FROM customers WHERE id = :id');
+        $STH->execute($data);
+        $message = 'Delete successful';
+    } else {
+        $STH = $DBH->prepare('UPDATE customers SET active = 1 WHERE id = :id');
+        $STH->execute($data);
+        $message = 'Cannot delete because of depency with 1 or more orders, the account has been disactivated';
+    }
 
     header('location:list.php?message=' . $message);
 } catch (PDOException $e) {
